HTML5 input type=number, and <h1> for the title
I'm using the HTML5 <input type=number> because it is semantically correct (it
is in fact a number), and because it provides a better UI (the proper number
keyboard on touch-screen; or the two arrow buttons when using a mouse).

I'm also using <h1> for the title, because that's the correct way to mark a
title.